New update: 17 Apr 2015
Read the top of the file "blcsz" found in /conky/bin  this is a complete
rewrite of the Zenity script that reads conky files automatically and
starts or stops each conky individually.

The conky files have been cleaned up, renamed and new ones added.
The conky: BL-Default.conkyrc is the new default conky for BunsenLabs

The README.md file here is updated as well.